First fully working icyparser release
Currently, icyparser works in a non-blocking way. That means, you can
call IcyParser(url).getIcyInformation(), and it will almost immediately
return a dictionary that contains the ICY information.

Every time that this is done, a connection will be made to the Icecast
server, requesting the information. Please don't be too agressive when
using icyparser and polling: short timeouts may or may not lead to a
DDOS-like effect on the server.

It's possible that I'll release a blocking icyparser in the future,
which may take some load off the server by connecting only once. Efforts
to do this have been made, but I'm having trouble finding a good way to
read out the bytestream yet not going out of bound.